Added TravisCI build
Next travis test

And the next test

Updated travis sudo access

Updated travis mkdir

Updated travis directories

Travis: Finding our home pt.1

Travis: Finding our gcc pt.2

Fixed bob bug

Travis: Finding our compiler pt.3

Try to fix scons by importing the PATH

Fixed strange bob behaviour

+ added ts3 travis notification

Revert "Try to fix scons by importing the PATH"

This reverts commit 1ce22390354fc669e9b8d99d1e8f1d6249f30d99.

Add CMake build system

Adjust travis.yml

Try to use clang on travis

Fix including header files for travis

Try to add more up to date compilers to travis

Next step, add swscale

+++ TS3AudioBot/TS3Query/Messages/ClientNotifications.cs
@@ -60,7 +60,7 @@
 		public bool IsRecording;
 
 		[QuerySerialized("client_database_id")]
-		public int DatabaseId;
+		public ulong DatabaseId;
 
 		[QuerySerialized("client_channel_group_id")]
 		public int ChannelGroupId;
